[#26] huge package refactor
