+++ src/main/java/com/beerjournal/breweriana/category/persistence/CategoryCrudRepository.java
@@ -1,0 +1,10 @@
+package com.beerjournal.breweriana.category.persistence;
+
+import org.bson.types.ObjectId;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import java.util.Optional;
+
+interface CategoryCrudRepository extends MongoRepository<Category, ObjectId> {
+    Optional<Category> findOneByName(String name);
+}
