+++ src/main/java/com/beerjournal/breweriana/collection/persistence/UserCollectionUpdater.java
@@ -1,0 +1,34 @@
+package com.beerjournal.breweriana.collection.persistence;
+
+import com.beerjournal.breweriana.item.persistence.Item;
+import com.beerjournal.breweriana.item.persistence.ItemUpdateListener;
+import com.beerjournal.breweriana.user.User;
+import com.beerjournal.breweriana.user.persistence.UserUpdateListener;
+import lombok.AllArgsConstructor;
+import org.bson.types.ObjectId;
+import org.springframework.stereotype.Component;
+
+import java.util.Collections;
+
+import static lombok.AccessLevel.PROTECTED;
+
+@Component
+@AllArgsConstructor(access = PROTECTED)
+class UserCollectionUpdater implements UserUpdateListener, ItemUpdateListener {
+
+    private UserCollectionRepository userCollectionRepository;
+
+    @Override
+    public void onInsert(User user) {
+        addEmptyCollection(user.getId());
+    }
+
+    @Override
+    public void onInsert(Item item) {
+        userCollectionRepository.addNewItem(item);
+    }
+
+    private UserCollection addEmptyCollection(ObjectId ownerId) {
+        return userCollectionRepository.save(UserCollection.of(ownerId, Collections.emptySet()));
+    }
+}
