+++ src/main/java/com/beerjournal/breweriana/item/ItemController.java
@@ -1,13 +1,10 @@
 package com.beerjournal.breweriana.item;
 
-import com.beerjournal.breweriana.collection.ItemRefDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.Collection;
-import java.util.Set;
 
 @RestController
 @RequestMapping("/api")
@@ -16,17 +13,15 @@ class ItemController {
 
     private final ItemService itemService;
 
-    @GetMapping("users/{userId}/collection/items")
-    ResponseEntity<Collection<ItemRefDto>> getAllItemRefsInUserCollection(
-            @RequestParam(value = "lacking", defaultValue = "false") boolean lacking,
-            @PathVariable(value = "userId") String id) {
-        Set<ItemRefDto> items = lacking ? itemService.getAllNotInUserCollection(id) : itemService.getAllItemRefsInUserCollection(id);
-        return new ResponseEntity<>(items, HttpStatus.OK);
-    }
-
     @GetMapping("items/{itemId}")
     ResponseEntity<ItemDto> getItemDetails(@PathVariable(value = "itemId") String id) {
         return new ResponseEntity<>(itemService.getItemDetails(id), HttpStatus.OK);
+    }
+
+    @PostMapping("users/{ownerId}/collection/items")
+    ResponseEntity<ItemDto> addItemToCollection(@PathVariable(value = "ownerId") String ownerId,
+                                                @RequestBody @Validated ItemDto item) {
+        return new ResponseEntity<>(itemService.addItem(ownerId, item), HttpStatus.CREATED);
     }
 
 }
