+++ src/main/java/com/beerjournal/breweriana/item/persistence/Attribute.java
@@ -1,0 +1,11 @@
+package com.beerjournal.breweriana.item.persistence;
+
+import lombok.Data;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+@Document
+@Data(staticConstructor = "of")
+public class Attribute {
+    private final String name;
+    private final Object value;
+}
