+++ src/main/java/com/beerjournal/breweriana/item/persistence/ItemUpdateListener.java
@@ -1,0 +1,11 @@
+package com.beerjournal.breweriana.item.persistence;
+
+public interface ItemUpdateListener {
+
+    default void onInsert(Item item) {}
+
+    default void onUpdate(Item item) {}
+
+    default void onDelete(Item item) {}
+
+}
