+++ src/main/java/com/beerjournal/breweriana/user/UserDto.java
@@ -1,6 +1,5 @@
 package com.beerjournal.breweriana.user;
 
-import com.beerjournal.breweriana.persistence.user.User;
 import lombok.Builder;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
