+++ src/main/java/com/beerjournal/breweriana/user/UserService.java
@@ -1,7 +1,6 @@
 package com.beerjournal.breweriana.user;
 
-import com.beerjournal.breweriana.persistence.UserRepository;
-import com.beerjournal.breweriana.persistence.user.User;
+import com.beerjournal.breweriana.user.persistence.UserRepository;
 import com.beerjournal.breweriana.utils.ServiceUtils;
 import com.beerjournal.infrastructure.error.BeerJournalException;
 import lombok.RequiredArgsConstructor;
