+++ src/main/java/com/beerjournal/breweriana/user/persistence/UserCrudRepository.java
@@ -1,0 +1,12 @@
+package com.beerjournal.breweriana.user.persistence;
+
+import com.beerjournal.breweriana.user.User;
+import org.bson.types.ObjectId;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import java.util.Optional;
+
+interface UserCrudRepository extends MongoRepository<User, ObjectId>{
+    Optional<User> findOneById(ObjectId id);
+    Optional<User> findOneByEmail(String email);
+}
