+++ src/main/java/com/beerjournal/breweriana/user/persistence/UserRepository.java
@@ -1,0 +1,39 @@
+package com.beerjournal.breweriana.user.persistence;
+
+import com.beerjournal.breweriana.user.User;
+import com.google.common.collect.ImmutableSet;
+import lombok.RequiredArgsConstructor;
+import org.bson.types.ObjectId;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+import java.util.Set;
+
+@Repository
+@RequiredArgsConstructor
+public class UserRepository {
+
+    private final UserCrudRepository crudRepository;
+    private final Set<UserUpdateListener> userUpdateListeners;
+
+    public ImmutableSet<User> findAll() {
+        return ImmutableSet.<User>builder()
+                .addAll(crudRepository.findAll())
+                .build();
+    }
+
+    public Optional<User> findOneById(ObjectId objectId) {
+        return crudRepository.findOneById(objectId);
+    }
+
+    public Optional<User> findOneByEmail(String email) {
+        return crudRepository.findOneByEmail(email);
+    }
+
+    public User save(User user) {
+        User savedUser = crudRepository.save(user);
+        userUpdateListeners.forEach(listener -> listener.onInsert(user));
+        return savedUser;
+    }
+
+}
