+++ src/main/java/com/beerjournal/infrastructure/security/BjPrincipal.java
@@ -1,6 +1,6 @@
 package com.beerjournal.infrastructure.security;
 
-import com.beerjournal.breweriana.persistence.user.User;
+import com.beerjournal.breweriana.user.User;
 import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
